Add borderconfig subcommand to stormloader
This subscommand is just a simple wrapper to put the configuration items
for the border router into the flash attributes, where the border router
code can retrieve them when it starts up. This command loads the
attributes into their expected locations and then launches the payload
to restart the mote

I was going to do some fancier encoding of the IPv4 addresses, but
because they're stored in bytes anyway, it didn't make much sense, so
that encoding is just done in the kernel. The keys here are just to
provide references for humans. The kernel knows which section should
contain which value.